+++ tests/NeuralNet/FeedForwardTest.php
@@ -86,7 +86,7 @@ class FeedForwardTest extends TestCase
      */
     public function layers() : void
     {
-        $this->assertCount(7, $this->network->layers());
+        $this->assertCount(7, iterator_to_array($this->network->layers()));
     }
 
     /**
@@ -111,6 +111,16 @@ class FeedForwardTest extends TestCase
     public function output() : void
     {
         $this->assertInstanceOf(Output::class, $this->network->output());
+    }
+
+    /**
+     * @test
+     */
+    public function numParams() : void
+    {
+        $this->network->initialize();
+
+        $this->assertEquals(103, $this->network->numParams());
     }
 
     /**